requirements.txt
django
requests
django-cors-headers
python-chess
pillow
django-ratelimit
python-dotenv
ndjson
openai>=1.0.0
stripe
redis
celery
pytest
pytest-django
pytest-cov
djangorestframework
djangorestframework-simplejwt
